Refactor: Organize Home screen components and states
This commit introduces several changes to enhance the structure and organization of the Home screen:

- **Component Refactoring:**
    - `AppInactiveScreen` is renamed to `InactiveAppHomeContent` and moved to the `content` package.
    - `AboutScreen` is moved from the `home` feature to a new `about` feature module.
    - New components are created for better modularity:
        - `CompactPager`: Manages the compact view of the Home screen with a pager for switching between main content and video player.
        - `HomePageButton`: A dedicated composable for the "Let's Go!" button.
        - `MainHomeContent`: Encapsulates the primary content of the Home screen, including the dancing bot and decorative squiggles.
        - `VideoPlayerRotatedCard`: Displays the video player within a rotated card.
- **State Management:**
    - `HomeState` data class is introduced to manage the state of the Home screen, including app active status, video link, and dancing droid link.
    - `HomeViewModel` is updated to use the new `HomeState` and initialize its values from `ConfigProvider`.
- **Utilities:**
    - `onVisibilityChanged` modifier is added to detect when a composable becomes visible within its container.
- **Navigation:**
    - `MainNavigation` is updated to reflect the new location of `AboutScreen`.
- **Code Organization:**
    - Existing Home screen composables are moved into more specific files within the `feature/home` module (e.g., `component`, `content`, `model`, `util`).
    - Imports for `ViewModelStoreNavEntryDecorator`, `SavedStateNavEntryDecorator`, and `SceneSetupNavEntryDecorator` are removed from `MainNavigation.kt` as they are no longer directly used.
    - `TransformOrigin` import removed from `MainNavigation.kt`.

These changes aim to improve code readability, maintainability, and scalability of the Home screen feature.
