+++ feature/home/src/main/java/com/android/developers/androidify/home/model/HomeState.kt
@@ -1,0 +1,7 @@
+package com.android.developers.androidify.home.model
+
+data class HomeState(
+    val isAppActive: Boolean = true,
+    val videoLink: String? = null,
+    val dancingDroidLink: String? = null,
+)
